0.1.0: union handler, refactor tests
