+++ package.json
@@ -1,13 +1,14 @@
 {
   "name": "@dotvirus/yxc",
-  "version": "0.0.7",
+  "version": "0.1.0",
   "description": "Schema validation with composition & API middlewares",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
+    "coverage": "nyc npm test",
     "build": "tsc",
-    "test:verbose": "set DEBUG=yxc && mocha --bail --exit -r ts-node/register test/index.ts",
-    "test": "mocha --bail --exit -r ts-node/register test/index.ts"
+    "test:verbose": "set DEBUG=yxc && mocha --exit -r ts-node/register test/**/spec.ts",
+    "test": "mocha --exit -r ts-node/register test/**/*.spec.ts"
   },
   "author": "dotvirus",
   "license": "MIT",
@@ -33,6 +34,7 @@
     "@types/node": "^14.0.1",
     "chai": "^4.2.0",
     "mocha": "^8.0.1",
+    "nyc": "^15.1.0",
     "ts-node": "^8.8.1",
     "typescript": "^3.8.3"
   }
