+++ src/index.ts
@@ -11,6 +11,7 @@ import { NumberHandler } from "./handlers/number";
 import { BooleanHandler } from "./handlers/boolean";
 import { ArrayHandler } from "./handlers/array";
 import { AnyHandler } from "./handlers/any";
+import { UnionHandler } from "./handlers/union";
 import { connect, koa, graphql } from "./middlewares";
 import { Handler } from "./handlers/index";
 
@@ -32,4 +33,5 @@ export default {
   boolean: () => new BooleanHandler(),
   array: <T = any>(handler: Handler) => new ArrayHandler<T>(handler),
   any: () => new AnyHandler(),
+  union: (handlers: Handler[]) => new UnionHandler(handlers),
 };
