--- test/index.ts
@@ -1,14 +1,0 @@
-import objectTests from "./object";
-objectTests();
-
-import stringTests from "./string";
-stringTests();
-
-import personSchemaTests from "./schemas/person";
-personSchemaTests();
-
-import musicSchemaTests from "./schemas/music";
-musicSchemaTests();
-
-import nestedSchemaTests from "./schemas/nested";
-nestedSchemaTests();
